requirements.txt
langchain
langchain-openai
openai
faiss-cpu
pandas
numpy
plotly
scikit-learn
python-dotenv
tiktoken



# streamlit==1.28.1
# openai==1.3.5
# langchain==0.0.327
# faiss-cpu==1.7.4
# pandas==2.1.3
# numpy==1.25.2
# plotly==5.17.0
# scikit-learn==1.3.2
# python-dotenv==1.0.0
# tiktoken==0.5.1
